Having some issues when I decrease the focal length down to 0.05, turns to a pink grey image instead of a bright color image.

+++ renderer/CoordinateSystem.cpp
@@ -11,9 +11,9 @@ CoordinateSystem::CoordinateSystem(sivelab::Vector3D direction, sivelab::Vector3
     direction = direction;
     W = direction;
     W *= (-1);
-    W /= direction.length();
+    W /= (float)direction.length();
     U = temp.cross(W);
-    U /=(temp.cross(W).length());
+    U /=(float)(temp.cross(W).length());
     V = W.cross(U);
 
 }
